changed short hand notation to long-hand old school notation
